+++ lib/src/json/commands/call_command.g.dart
@@ -8,13 +8,15 @@ part of 'call_command.dart';
 
 CallCommand _$CallCommandFromJson(Map<String, dynamic> json) => CallCommand(
       commandId: json['commandId'] as String,
+      conditions: (json['conditions'] as List<dynamic>?)
+          ?.map((e) => Conditional.fromJson(e as Map<String, dynamic>))
+          .toList(),
       callAfter: json['callAfter'] as int?,
-      chance: json['chance'] as int? ?? 1,
     );
 
 Map<String, dynamic> _$CallCommandToJson(CallCommand instance) =>
     <String, dynamic>{
+      'conditions': instance.conditions,
       'commandId': instance.commandId,
       'callAfter': instance.callAfter,
-      'chance': instance.chance,
     };